Added middlewares to manage token Auth

+++ controllers/blogs.js
@@ -1,6 +1,8 @@
 import express from 'express';
+
 import Blog from '../models/blog';
 import User from '../models/user';
+import { userExtractor } from '../utils/middleware';
 
 const blogsRouter = express.Router();
 
@@ -13,11 +15,11 @@ blogsRouter.get('/', async (req, res) => {
   res.json(blogs);
 });
 
-blogsRouter.post('/', async (req, res) => {
-  const { title, author, url, likes, userId } = req.body;
+blogsRouter.post('/', userExtractor, async (req, res) => {
+  const { title, author, url, likes } = req.body;
 
   // Find user
-  const user = await User.findById(userId);
+  const user = await User.findById(req.user.id);
 
   if (!user) {
     return res.status(400).json({ error: 'User not found' });
@@ -29,7 +31,6 @@ blogsRouter.post('/', async (req, res) => {
     url,
     likes: likes || 0,
     date: new Date(),
-    // eslint-disable-next-line no-underscore-dangle
     user: user.id,
   });
 
@@ -49,7 +50,7 @@ blogsRouter.get('/:id', async (req, res) => {
   }
 });
 
-blogsRouter.put('/:id', async (req, res) => {
+blogsRouter.put('/:id', userExtractor, async (req, res) => {
   const { title, author, url, likes, date } = req.body;
 
   const blog = {
@@ -66,9 +67,16 @@ blogsRouter.put('/:id', async (req, res) => {
   res.json(updatedBlog.toJSON());
 });
 
-blogsRouter.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', userExtractor, async (req, res) => {
+  // Check if the blog owner is the same as the user making the request
+  const blog = await Blog.findById(req.params.id);
+
+  if (blog.user.toString() !== req.user.id.toString()) {
+    return res.status(401).json({ error: 'Unauthorized' });
+  }
+
   await Blog.findByIdAndRemove(req.params.id);
-  res.status(204).end();
+  return res.status(204).end();
 });
 
 export default blogsRouter;
